implementa css ao footer e implementa separacao de pasta
