+++ src/components/FooterMenu/FooterMenu.css
@@ -1,0 +1,11 @@
+footer {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+}
+
+.buttons-footer {
+  display: flex;
+  justify-content: space-around;
+}
